adding favorites screen + moving functions to SongDAO.kt
