+++ app/src/main/java/com/example/songhub/DAO/SongDAO.kt
@@ -2,9 +2,16 @@ package com.example.songhub.DAO
 
 import android.net.Uri
 import com.example.songhub.model.Song
+import com.github.kittinunf.fuel.Fuel
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.toObjects
 import com.google.firebase.storage.FirebaseStorage
+import com.google.gson.Gson
+import com.google.gson.JsonObject
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 
 class SongDAO {
     val db = FirebaseFirestore.getInstance()
@@ -75,6 +82,7 @@ class SongDAO {
         } ?: callback(false)
     }
 
+
     fun findById(id: String, callback: (Song?) -> Unit) {
         db.collection("songs").document(id).get()
             .addOnSuccessListener { document ->
@@ -90,4 +98,68 @@ class SongDAO {
                 callback(null)
             }
     }
+
+    private fun extractArtistAndTrack(url: String): Pair<String, String> {
+        val parts = url.split("/_/")
+        if (parts.size == 2) {
+            val artist = parts[0].substringAfterLast("/").replace("+", "%20")
+            val track = parts[1].replace("+", "%20")
+            println("AQUI ESTÃO: $artist e o $track")
+            return artist to track
+        }
+        throw IllegalArgumentException("URL malformada: $url")
+    }
+
+    private fun getTrackInfo(artist: String, track: String, apiKey: String): JsonObject? {
+        val url = "https://ws.audioscrobbler.com/2.0/?method=track.getInfo&artist=$artist&track=$track&api_key=$apiKey&format=json"
+
+        val (request, response, result) = Fuel.get(url).responseString()
+
+        return when (result) {
+            is com.github.kittinunf.result.Result.Success -> {
+                val gson = Gson()
+                gson.fromJson(result.value, JsonObject::class.java)
+            }
+            is com.github.kittinunf.result.Result.Failure -> {
+                println("Erro na requisição: ${result.getException()}")
+                null
+            }
+        }
+    }
+
+    fun fetchTracksInfo(urls: List<String>, apiKey: String, callback: (List<Song>) -> Unit) {
+        CoroutineScope(Dispatchers.IO).launch {
+            val results = urls.map { url ->
+                val (artist, track) = extractArtistAndTrack(url)
+                getTrackInfo(artist, track, apiKey)
+            }
+
+            // Map the JsonObject results to Song objects
+            val songs = results.mapNotNull { trackInfo ->
+                trackInfo?.let { info ->
+                    val track = info.getAsJsonObject("track")
+                    val artistName = track.getAsJsonObject("artist").get("name").asString
+                    val trackName = track.get("name").asString
+                    val url = track.get("url").asString
+
+                    // Extract the image URL from the album object
+                    val imagesArray = track.getAsJsonObject("album").getAsJsonArray("image")
+                    val imageUrl = if (imagesArray.size() > 3) {
+                        imagesArray[3].asJsonObject.get("#text").asString
+                    } else {
+                        null // Handle the case where the index doesn't exist
+                    }
+
+                    // Assuming your Song model has these fields
+                    Song(title = trackName, artist = artistName, imageUrl = imageUrl, url = url)
+                }
+            }
+
+            // Switch to the main thread to invoke the callback with the list of songs
+            withContext(Dispatchers.Main) {
+                callback(songs)
+            }
+        }
+    }
+
 }
