+++ app/src/main/java/com/example/songhub/ui/screens/MainScreen.kt
@@ -24,6 +24,7 @@ import androidx.compose.ui.unit.sp
 import androidx.compose.ui.draw.clip
 import androidx.navigation.NavController
 import coil.compose.rememberImagePainter
+import com.example.songhub.DAO.SongDAO
 import com.example.songhub.DAO.UserDAO
 import com.example.songhub.R
 import com.example.songhub.model.Song
@@ -43,13 +44,14 @@ fun MainScreen(modifier: Modifier = Modifier, navController: NavController) {
     val coroutineScope = rememberCoroutineScope()
     var user = UserSession.loggedInUser
     var userDAO = UserDAO()
+    var songDAO = SongDAO()
     var songs = remember { mutableStateOf<List<Song>>(emptyList()) }
 
     LaunchedEffect(Unit) {
         if (user != null) {
             userDAO.getMySongs(user.username) { mySongs ->
                 if (mySongs != null && mySongs.isNotEmpty()) {
-                    fetchTracksInfo(mySongs, "499a9407d353802f5f07166c0d8f35c2") { fetchedSongs ->
+                    songDAO.fetchTracksInfo(mySongs, "499a9407d353802f5f07166c0d8f35c2") { fetchedSongs ->
                         // Update your songs state with the fetched songs
                         songs.value = fetchedSongs
                     }
@@ -84,69 +86,6 @@ fun MainScreen(modifier: Modifier = Modifier, navController: NavController) {
                 tint = Color(0xFF040723),
                 modifier = Modifier.size(30.dp)
             )
-        }
-    }
-}
-
-fun extractArtistAndTrack(url: String): Pair<String, String> {
-    val parts = url.split("/_/")
-    if (parts.size == 2) {
-        val artist = parts[0].substringAfterLast("/").replace("+", "%20")
-        val track = parts[1].replace("+", "%20")
-        println("AQUI ESTÃO: $artist e o $track")
-        return artist to track
-    }
-    throw IllegalArgumentException("URL malformada: $url")
-}
-
-fun getTrackInfo(artist: String, track: String, apiKey: String): JsonObject? {
-    val url = "https://ws.audioscrobbler.com/2.0/?method=track.getInfo&artist=$artist&track=$track&api_key=$apiKey&format=json"
-
-    val (request, response, result) = Fuel.get(url).responseString()
-
-    return when (result) {
-        is com.github.kittinunf.result.Result.Success -> {
-            val gson = Gson()
-            gson.fromJson(result.value, JsonObject::class.java)
-        }
-        is com.github.kittinunf.result.Result.Failure -> {
-            println("Erro na requisição: ${result.getException()}")
-            null
-        }
-    }
-}
-
-fun fetchTracksInfo(urls: List<String>, apiKey: String, callback: (List<Song>) -> Unit) {
-    CoroutineScope(Dispatchers.IO).launch {
-        val results = urls.map { url ->
-            val (artist, track) = extractArtistAndTrack(url)
-            getTrackInfo(artist, track, apiKey)
-        }
-
-        // Map the JsonObject results to Song objects
-        val songs = results.mapNotNull { trackInfo ->
-            trackInfo?.let { info ->
-                val track = info.getAsJsonObject("track")
-                val artistName = track.getAsJsonObject("artist").get("name").asString
-                val trackName = track.get("name").asString
-                val url = track.get("url").asString
-
-                // Extract the image URL from the album object
-                val imagesArray = track.getAsJsonObject("album").getAsJsonArray("image")
-                val imageUrl = if (imagesArray.size() > 3) {
-                    imagesArray[3].asJsonObject.get("#text").asString
-                } else {
-                    null // Handle the case where the index doesn't exist
-                }
-
-                // Assuming your Song model has these fields
-                Song(title = trackName, artist = artistName, imageUrl = imageUrl, url = url)
-            }
-        }
-
-        // Switch to the main thread to invoke the callback with the list of songs
-        withContext(Dispatchers.Main) {
-            callback(songs)
         }
     }
 }
